+++ src/test/java/by/pizzatempo/pages/MenuPage.java
@@ -1,4 +1,4 @@
-package pages;
+package by.pizzatempo.pages;
 
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
@@ -9,16 +9,19 @@ import org.openqa.selenium.support.ui.WebDriverWait;
 
 import java.time.Duration;
 
-public class MainPage {
+public class MenuPage {
     private WebDriverWait wait;
 
     @FindBy(xpath = "//a[@href='https://www.pizzatempo.by/menu/pizza.html']")
     WebElement pizzaPageButton;
 
+    @FindBy(xpath = "//a[@href='https://www.pizzatempo.by/menu/napitki.html']")
+    WebElement drinkPageButton;
+
     @FindBy(xpath = "//a[@href='https://www.pizzatempo.by/menu/order/']")
     WebElement basketPageButton;
 
-    public MainPage(WebDriver driver) {
+    public MenuPage(WebDriver driver) {
         this.wait = new WebDriverWait(driver, Duration.ofSeconds(5));
         PageFactory.initElements(driver, this);
     }
@@ -26,6 +29,11 @@ public class MainPage {
     public void clickPizzaButton() {
         wait.until(ExpectedConditions.elementToBeClickable(pizzaPageButton));
         pizzaPageButton.click();
+    }
+
+    public void clickDrinkButton() {
+        wait.until(ExpectedConditions.elementToBeClickable(drinkPageButton));
+        drinkPageButton.click();
     }
 
     public void clickBasketPageButton() {
