+++ src/test/java/by/pizzatempo/pages/OrderPopUp.java
@@ -1,4 +1,4 @@
-package pages;
+package by.pizzatempo.pages;
 
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
