+++ src/test/java/by/pizzatempo/tests/BaseTest.java
@@ -1,4 +1,4 @@
-package tests;
+package by.pizzatempo.tests;
 
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
